# src/docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "An E-commerce store with Express"
  },
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Get all products",
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "thumbnail": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "stock": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new product",
        "operationId": "createProduct",
        "requestBody": {
          "description": "Product to add to the store",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "thumbnail": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get one product",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Modify one product",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete one product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "thumbnail": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "stock": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carts/{id}/products/{idProd}": {
      "delete": {
        "summary": "Delete one product of a cart",
        "operationId": "deleteCartProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id_prod",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add one product of a cart",
        "operationId": "addCartProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id_prod",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get products of a cart",
        "operationId": "getCartProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id_prod",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}